Check if commands.json exists before trying to delete it

+++ src/Digital/Console/Generate.php
@@ -48,8 +48,10 @@ class Generate extends Command
 
         $output->writeln($process->getOutput());
 
-        $process = new Process('rm output/commands.json');
-        $process->run();
+        if (file_exists('output/commands.json')) {
+          $process = new Process('rm output/commands.json');
+          $process->run();
+        }
 
         // executes after the command finishes
         if (!$process->isSuccessful()) {
